backend/requirements.txt
celery==5.5.1
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
psycopg[binary]==3.2.6
redis==6.0.0
stripe==12.0.0
gunicorn==23.0.0
Pillow==11.2.1
